Move enumerate_hashless into `enumerate`
